requirements.txt
datasets
gradio
langchain-community
langchain_experimental
langchain_huggingface
langchain_google_genai
langchain_ollama
langgraph
python-dotenv
rank_bm25